requirements.txt
Zenora==0.0.3.post1
Flask==3.0.0
Flask-Login==0.6.3
python-dotenv==1.0.0
psycopg2==2.9.9
flask_sqlalchemy==3.1.1
disnake==2.9.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23